:recycle: Decompose SoundBiteGenerator into 9 focused modules (closes #968)
Splits the 852-line sound-bite-generator.ts into focused collaborators
with clear single responsibilities, each independently testable:

- sound-bite-lm-adapter.ts — stream collection, retry gate, content-policy detection
- response-parser.ts — JSON cleaning, shape validation, VoiceProfile coercion + error classes
- prompt-builders.ts — entity and scholar prompt builders + audio tag constants
- voice-mapping.ts — buildGeminiVoiceName (28-voice table) + buildVoiceStyleInstruction
- tts/gemini-tts-service.ts — Gemini 2.5 Flash TTS HTTP call + base64 decode
- tts/web-speech-tts-service.ts — browser SpeechSynthesis fallback
- tts/cascading-tts-service.ts — TTSService interface + cascade orchestration
- tts/pcm-to-wav.ts — pure RIFF/WAV encoder (testable without fetch mock)

Code-review fixes applied in the same pass:
- Retry now gates only on SoundBiteContentPolicyError (not all SoundBiteGenerationError),
  so prose-wrapped JSON responses correctly reach the strict-JSON retry prompt
- Enum validator arrays use `satisfies ReadonlyArray<VoiceProfile[...]>` so the
  compiler catches schema type drift
- _contextRetrieval made optional in constructor and factory (stored but not yet used)
- GeminiTTSService and WebSpeechTTSService restored to `implements TTSService`
- output variable typed as GenerationOutput directly instead of ReturnType<typeof>

All 25 existing tests pass unchanged. Public oracle-engine API surface is preserved.
