+++ packages/oracle-engine/src/tts/pcm-to-wav.ts
@@ -1,0 +1,35 @@
+/**
+ * Wraps raw 16-bit signed little-endian mono PCM bytes in a RIFF/WAV container.
+ * Gemini TTS always returns this format (audio/pcm;rate=24000).
+ */
+export function pcmBytesToWav(
+  pcm: Uint8Array,
+  sampleRate: number,
+  numChannels = 1,
+  bitsPerSample = 16,
+): Blob {
+  const byteRate = sampleRate * numChannels * (bitsPerSample / 8);
+  const blockAlign = numChannels * (bitsPerSample / 8);
+  const dataSize = pcm.byteLength;
+  const buffer = new ArrayBuffer(44 + dataSize);
+  const view = new DataView(buffer);
+  const write = (offset: number, str: string) => {
+    for (let i = 0; i < str.length; i++)
+      view.setUint8(offset + i, str.charCodeAt(i));
+  };
+  write(0, "RIFF");
+  view.setUint32(4, 36 + dataSize, true);
+  write(8, "WAVE");
+  write(12, "fmt ");
+  view.setUint32(16, 16, true);
+  view.setUint16(20, 1, true);
+  view.setUint16(22, numChannels, true);
+  view.setUint32(24, sampleRate, true);
+  view.setUint32(28, byteRate, true);
+  view.setUint16(32, blockAlign, true);
+  view.setUint16(34, bitsPerSample, true);
+  write(36, "data");
+  view.setUint32(40, dataSize, true);
+  new Uint8Array(buffer, 44).set(pcm);
+  return new Blob([buffer], { type: "audio/wav" });
+}
